+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.3] - 2025-05-05
+
+### Changed
+
+- Added default latitude and longitude to the config;
+- Added option to set `GOOGLE_MAPS_API_KEY` in the environment variables;
+- Replaced `axios` with `useFetchClient`;
+
 ## [1.0.0] - 2025-05-03
 
 ### Changed